+++ cdn/locals.tf
@@ -1,3 +1,0 @@
-locals {
-  s3_origin_id = "S3-${var.domain_name}"
-}