+++ s3-private/outputs.tf
@@ -1,0 +1,7 @@
+output "arn" {
+  value = aws_s3_bucket.main.arn
+}
+
+output "kms-key-arn" {
+  value = module.kms-key-s3.arn
+}