+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
 
+## Unreleased
+
+### RDS
+- an initial database is now created with the name `project_environment` [#52](https://github.com/dbl-works/terraform/pull/52)
+
+
 
 ## [v2021.11.13] - 2021-11-13