+++ CHANGELOG.md
@@ -9,9 +9,15 @@ This project does not follow SemVer, since modules are independent of each other
 
 ## Unreleased
 
+### RDS
+
+- an initial database is now created with the name `{project}_{environment}` [#52](https://github.com/dbl-works/terraform/pull/52)
+
 ### Added
 
 - new module `s3-private`, a basic S3 bucket for private, encrypted files [#51](https://github.com/dbl-works/terraform/pull/51)
+
+
 
 ### CDN